backend/requirements.txt
flask
celery
redis
yt-dlp
google-generativeai
google-cloud-firestore
Flask-Cors
gunicorn